end line appends removed, db exception handled, readme updated

+++ src/test/java/fatjar/test/WebTester.java
@@ -44,36 +44,40 @@ public class WebTester {
                 })
                 .get("/db", (req, res) -> {
 
-                    DB db = DB.create();
+                    Optional<DB> dbOptional = DB.create();
+                    if (dbOptional.isPresent()) {
+                        DB db = dbOptional.get();
+                        long numberOfAllEntities = db.count(MyEntity.class);
+                        long numberOfJohnEntities = db.count(MyEntity.class, DB.Query.create("name", "five"));
+                        long numberOfLargeIDEntities = db.count(MyEntity.class, DB.Query.create("id", DB.Sign.GT, 1L));
+                        Log.info("numberOfLargeIDEntities: " + numberOfLargeIDEntities);
 
-                    long numberOfAllEntities = db.count(MyEntity.class);
-                    long numberOfJohnEntities = db.count(MyEntity.class, DB.Query.create("name", "five"));
-                    long numberOfLargeIDEntities = db.count(MyEntity.class, DB.Query.create("id", DB.Sign.GT, 1L));
-                    Log.info("numberOfLargeIDEntities: " + numberOfLargeIDEntities);
+                        List<MyEntity> allEntities = db.findAll(MyEntity.class);
+                        Log.info("found allEntities: " + allEntities);
 
-                    List<MyEntity> allEntities = db.findAll(MyEntity.class);
-                    Log.info("found allEntities: " + allEntities);
+                        MyEntity entityID1L = db.find(MyEntity.class, 1L);
+                        Log.info("fount entityID1L: " + entityID1L);
 
-                    MyEntity entityID1L = db.find(MyEntity.class, 1L);
-                    Log.info("fount entityID1L: " + entityID1L);
+                        MyEntity entity = new MyEntity("john");
+                        Optional<MyEntity> insertedEntityOptional = db.insert(entity);
+                        if (insertedEntityOptional.isPresent()) {
+                            MyEntity insertedEntity = insertedEntityOptional.get();
+                            Log.info("insert insertedEntity: " + insertedEntity);
 
-                    MyEntity entity = new MyEntity("john");
-                    Optional<MyEntity> insertedEntityOptional = db.insert(entity);
-                    if (insertedEntityOptional.isPresent()) {
-                        MyEntity insertedEntity = insertedEntityOptional.get();
-                        Log.info("insert insertedEntity: " + insertedEntity);
+                            insertedEntity.setName("johnny");
+                            db.update(insertedEntity);
+                            Log.info("update insertedEntity: " + insertedEntity);
 
-                        insertedEntity.setName("johnny");
-                        db.update(insertedEntity);
-                        Log.info("update insertedEntity: " + insertedEntity);
-
-                        db.delete(insertedEntity);
-                        Log.info("delete insertedEntity: " + insertedEntity);
-                        res.setContent(JSON.toJson(insertedEntity));
+                            db.delete(insertedEntity);
+                            Log.info("delete insertedEntity: " + insertedEntity);
+                            res.setContent(JSON.toJson(insertedEntity));
+                        } else {
+                            res.setContent("could not insert to DB");
+                        }
+                        res.write();
                     } else {
-                        res.setContent("could not insert to DB");
+                        throw new Server.ServerException("no db available");
                     }
-                    res.write();
                 })
                 .get("/httpClient", (req, res) -> {
                     try {
